Refactor app.py to improve code structure

+++ app.py
@@ -119,10 +119,10 @@ if selected == "Heart Disease Prediction":
             st.warning("⚠️ Please fill in all fields")
         else:
             try:
-                data = np.array([[int(age), int(sex), int(cp), float(trestbps),
-                                  float(chol), int(fbs), int(restecg),
-                                  float(thalach), int(exang), float(oldpeak),
-                                  int(slope), int(ca), int(thal)]])
+               data = np.array([[float(age), float(sex), float(cp), float(trestbps),
+                    float(chol), float(fbs), float(restecg),
+                    float(thalach), float(exang), float(oldpeak),
+                    float(slope), float(ca), float(thal)]])
 
                 prediction = heart_disease_model.predict(data)
 
@@ -208,6 +208,7 @@ if selected == "Breast Cancer Prediction":
                     st.success("✅ Benign Tumor")
             except:
                 st.error("❌ Invalid numeric input")
+
